+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [x]
+- Course management: User language column and filter
+
 ## [2.3.0]
 - Enable users and courses view config
 - Course administration (Enrolment from reporting view)